+++ src/output/TextTriggerTrait.php
@@ -1,7 +1,7 @@
 <?php
 namespace phphound\output;
 
-use phphound\Command;
+use phphound\Analyser;
 
 /**
  * Outputs events information to the console.
@@ -18,28 +18,25 @@ trait TextTriggerTrait
     public function trigger($eventType, $message = null)
     {
         switch ($eventType) {
-            case Command::EVENT_STARTING_ANALYSIS:
+            case Analyser::EVENT_STARTING_ANALYSIS:
                 $this->cli->green('Starting analysis');
-                break;
-
-            case Command::EVENT_STARTING_TOOL:
-                $this->cli->inline('Running ' . $message['description'] . '... ');
                 if (!empty($message['ignoredPaths'])) {
-                    $this->cli->br();
-                    $this->cli->inline('Ignored paths:');
-                    $this->cli->br();
+                    $this->cli->out('(ignored paths:)');
                     foreach ($message['ignoredPaths'] as $ignoredPath) {
-                        $this->cli->inline('     ' . $ignoredPath);
-                        $this->cli->br();
+                        $this->cli->red("\t" . $ignoredPath);
                     }
                 }
                 break;
 
-            case Command::EVENT_FINISHED_TOOL:
+            case Analyser::EVENT_STARTING_TOOL:
+                $this->cli->inline('Running ' . $message['description'] . '... ');
+                break;
+
+            case Analyser::EVENT_FINISHED_TOOL:
                 $this->cli->out('Done!');
                 break;
 
-            case Command::EVENT_FINISHED_ANALYSIS:
+            case Analyser::EVENT_FINISHED_ANALYSIS:
                 $this->cli->br();
                 $this->cli->green('Analysis complete!');
                 break;