+++ tests/output/HtmlOutputTest.php
@@ -1,7 +1,7 @@
 <?php
 namespace tests\output;
 
-use phphound\Command;
+use phphound\Analyser;
 use phphound\output\HtmlOutput;
 
 class HtmlOutputTest extends \PHPUnit_Framework_TestCase
@@ -58,14 +58,20 @@ class HtmlOutputTest extends \PHPUnit_Framework_TestCase
     public function it_outputs_on_starting_analysis()
     {
         $cli = $this->getMockBuilder('League\CLImate\CLImate')
-            ->setMethods(['green'])
+            ->setMethods(['green', 'red', 'out'])
             ->getMock()
         ;
         $output = new HtmlOutput($cli, $this->runningScriptDir);
 
-        $cli->expects($this->once())->method('green')->with('Starting analysis');
+        $cli->expects($this->at(0))->method('green')->with('Starting analysis');
+        $cli->expects($this->at(1))->method('out')->with('(ignored paths:)');
+        $cli->expects($this->at(2))->method('red')->with("\tvendor");
+        $cli->expects($this->at(3))->method('red')->with("\ttests");
 
-        $output->trigger(Command::EVENT_STARTING_ANALYSIS);
+        $output->trigger(
+            Analyser::EVENT_STARTING_ANALYSIS,
+            ['ignoredPaths' => ['vendor', 'tests']]
+        );
     }
 
     /** @test */
@@ -76,14 +82,11 @@ class HtmlOutputTest extends \PHPUnit_Framework_TestCase
             ->getMock()
         ;
         $output = new HtmlOutput($cli, $this->runningScriptDir);
-        $message = ['description' => 'Toolname', 'ignoredPaths' => ['vendor', 'tests']];
-        $cli->expects($this->at(0))->method('inline')->with('Running Toolname... ');
+        $message = ['description' => 'Toolname'];
 
-        $cli->expects($this->at(1))->method('inline')->with('Ignored paths:');
-        $cli->expects($this->at(2))->method('inline')->with('     ' . $message['ignoredPaths'][0]);
-        $cli->expects($this->at(3))->method('inline')->with('     ' . $message['ignoredPaths'][1]);
+        $cli->expects($this->once())->method('inline')->with('Running Toolname... ');
 
-        $output->trigger(Command::EVENT_STARTING_TOOL, $message);
+        $output->trigger(Analyser::EVENT_STARTING_TOOL, $message);
     }
 
     /** @test */
@@ -97,7 +100,7 @@ class HtmlOutputTest extends \PHPUnit_Framework_TestCase
 
         $cli->expects($this->once())->method('out')->with('Done!');
 
-        $output->trigger(Command::EVENT_FINISHED_TOOL);
+        $output->trigger(Analyser::EVENT_FINISHED_TOOL);
     }
 
     /** @test */
@@ -111,6 +114,6 @@ class HtmlOutputTest extends \PHPUnit_Framework_TestCase
 
         $cli->expects($this->once())->method('green')->with('Analysis complete!');
 
-        $output->trigger(Command::EVENT_FINISHED_ANALYSIS);
+        $output->trigger(Analyser::EVENT_FINISHED_ANALYSIS);
     }
 }