+++ CHANGELOG.md
@@ -5,9 +5,11 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## Unreleased
 ### Added
 - Add a history chart to HTML output, displaying quality issues over time.
+- Output a list of the ignored paths.
 
 ### Changed
-- Each third party tool generates its own `AnalysisResult`
+- Each third party tool generates its own `AnalysisResult`.
+- Extracts an `Analyser` class from `Command`.
 
 ## 0.5.1 - 2015-08-19
 ### Fixed